requirements.txt
numpy==1.17.2
matplotlib==3.1.1
Pillow==6.2.0
requests==2.22.0